polymarket/requirements.txt
requests>=2.31
pandas>=2.0
python-dotenv>=1.0
loguru>=0.7
matplotlib>=3.7
pytest>=7.4
